updating arm.py to handle multiple instances in a single process

+++ interbotix_xs_toolbox/interbotix_xs_modules/interbotix_xs_modules/xs_robot/arm.py
@@ -37,6 +37,7 @@ from threading import Thread
 import time
 from typing import Any, List, Tuple, Union
 
+from builtin_interfaces.msg import Duration
 import interbotix_common_modules.angle_manipulation as ang
 from interbotix_xs_modules.xs_robot import mr_descriptions as mrd
 from interbotix_xs_modules.xs_robot.core import InterbotixRobotXSCore
@@ -51,7 +52,6 @@ import modern_robotics as mr
 import numpy as np
 import rclpy
 from rclpy.constants import S_TO_NS
-from rclpy.duration import Duration
 from rclpy.executors import MultiThreadedExecutor
 from rclpy.logging import LoggingSeverity
 from trajectory_msgs.msg import JointTrajectory, JointTrajectoryPoint
@@ -79,6 +79,8 @@ class InterbotixManipulatorXS:
         logging_level: LoggingSeverity = LoggingSeverity.INFO,
         node_name: str = 'robot_manipulation',
         start_on_init: bool = True,
+        node_owner: bool = True,
+        update_kinematics: bool = True,
         args=None,
     ) -> None:
         """
@@ -113,6 +115,12 @@ class InterbotixManipulatorXS:
             ERROR, or FATAL. defaults to INFO
         :param node_name: (optional) name to give to the core started by this class, defaults to
             'robot_manipulation'
+        :param node_owner: (Used when controlling multiple robots) set to `True` to give access
+            to rclpy states like init and shutdown. For multiple robots, only one robot should
+            get the node ownership. Every other robot will use the node but cannot control init
+            or shutdown.
+        :param update_kinematics: (optional) decides whether or not to update forward
+            kinematics of the arm. Can be disabled to save computations.
         :param start_on_init: (optional) set to `True` to start running the spin thread after the
             object is built; set to `False` if intending to sub-class this. If set to `False`,
             either call the `start()` method later on, or add the core to an executor in another
@@ -132,6 +140,7 @@ class InterbotixManipulatorXS:
             group_name=group_name,
             moving_time=moving_time,
             accel_time=accel_time,
+            update_kinematics=update_kinematics
         )
         if gripper_name is not None:
             self.gripper = InterbotixGripperXSInterface(
@@ -142,7 +151,10 @@ class InterbotixManipulatorXS:
                 gripper_pressure_upper_limit=gripper_pressure_upper_limit,
             )
 
-        if start_on_init:
+        self.node_owner = node_owner
+
+        if start_on_init and self.node_owner:
+            self.core.get_logger().info("Starting thread for owner")
             self.start()
 
     def start(self) -> None:
@@ -158,10 +170,15 @@ class InterbotixManipulatorXS:
 
     def shutdown(self) -> None:
         """Destroy the node and shut down all threads and processes."""
-        self.core.destroy_node()
-        rclpy.shutdown()
-        self._execution_thread.join()
-        time.sleep(0.5)
+        if self.node_owner:
+            self.core.destroy_node()
+            rclpy.shutdown()
+            self._execution_thread.join()
+            time.sleep(0.5)
+        else:
+            self.core.get_logger().error(
+                "Cannot perform shutdown due to lack of ownership"
+            )
 
 
 class InterbotixArmXSInterface:
@@ -174,6 +191,7 @@ class InterbotixArmXSInterface:
         group_name: str,
         moving_time: float = 2.0,
         accel_time: float = 0.3,
+        update_kinematics: bool = True
     ) -> None:
         """
         Construct the InterbotixArmXSInterface object.
@@ -187,11 +205,14 @@ class InterbotixArmXSInterface:
             complete one move
         :param accel_time: (optional) time [s] it should take for all joints in the arm to
             accelerate/decelerate to/from max speed
+        :param update_kinematics: (optional) decides whether or not to update forward
+            kinematics of the arm. Can be disabled to save computations.
         """
         self.core = core
         self.robot_model = robot_model
         self.moving_time, self.accel_time = moving_time, accel_time
         self.group_name = group_name
+        self.update_kinematics = update_kinematics
 
         self.robot_des: mrd.ModernRoboticsDescription = getattr(mrd, self.robot_model)
 
@@ -214,8 +235,7 @@ class InterbotixArmXSInterface:
             )
             exit(1)
 
-        # initialize initial IK guesses
-        self.initial_guesses = [[0.0] * self.group_info.num_joints for _ in range(3)]
+        self.initial_guesses = [[0.0] * self.group_info.num_joints] * 3
         self.initial_guesses[1][0] = np.deg2rad(-120)
         self.initial_guesses[2][0] = np.deg2rad(120)
         self.joint_commands = []
@@ -270,8 +290,11 @@ class InterbotixArmXSInterface:
         )
         self.core.pub_group.publish(joint_commands)
         if blocking:
-            self.core.get_clock().sleep_for(Duration(nanoseconds=int(self.moving_time*S_TO_NS)))
-        self._update_Tsb()
+            time.sleep(
+                self.moving_time
+            )  # TODO: once released, use rclpy.clock().sleep_for()
+        if self.update_kinematics:
+            self._update_Tsb()
 
     def set_trajectory_time(
         self,
@@ -298,7 +321,10 @@ class InterbotixArmXSInterface:
                     value=int(moving_time * 1000),
                 )
             )
-            self.core.robot_spin_until_future_complete(future_moving_time)
+            self.core.executor.spin_once_until_future_complete(
+                future=future_moving_time,
+                timeout_sec=0.1
+            )
 
         if accel_time is not None and accel_time != self.accel_time:
             self.accel_time = accel_time
@@ -310,7 +336,10 @@ class InterbotixArmXSInterface:
                     value=int(accel_time * 1000),
                 )
             )
-            self.core.robot_spin_until_future_complete(future_accel_time)
+            self.core.executor.spin_once_until_future_complete(
+                future=future_accel_time,
+                timeout_sec=0.1
+            )
 
     def _check_joint_limits(self, positions: List[float]) -> bool:
         """
@@ -464,7 +493,7 @@ class InterbotixArmXSInterface:
         single_command = JointSingleCommand(name=joint_name, cmd=position)
         self.core.pub_single.publish(single_command)
         if blocking:
-            self.core.get_clock().sleep_for(Duration(nanoseconds=int(self.moving_time*S_TO_NS)))
+            time.sleep(self.moving_time)
         self._update_Tsb()
         return True
 
@@ -656,8 +685,8 @@ class InterbotixArmXSInterface:
             joint_traj_point = JointTrajectoryPoint()
             joint_traj_point.positions = tuple(joint_positions)
             joint_traj_point.time_from_start = Duration(
-                nanoseconds=int(i * wp_period * S_TO_NS)
-            ).to_msg()
+                nanosec=int(i * wp_period * S_TO_NS)
+            )
             joint_traj.points.append(joint_traj_point)
             if i == N:
                 break
@@ -697,9 +726,7 @@ class InterbotixArmXSInterface:
                     cmd_type='group', name=self.group_name, traj=joint_traj
                 )
             )
-            self.core.get_clock().sleep_for(
-                Duration(nanoseconds=int((moving_time + wp_moving_time)*S_TO_NS))
-            )
+            time.sleep(moving_time + wp_moving_time)
             self.T_sb = T_sd
             self.joint_commands = joint_positions
             self.set_trajectory_time(moving_time, accel_time)
